ml-engine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.3
scikit-learn>=1.3.2
umap-learn>=0.5.5
sentence-transformers>=2.2.2
joblib>=1.3.2
python-multipart>=0.0.6
python-dotenv>=1.0.0

